fix hang
