+++ oom_tracker_darwin.go
@@ -3,12 +3,12 @@
 package jasper
 
 import (
-	"bufio"
 	"context"
 	"os/exec"
+	"regexp"
+	"strconv"
 	"strings"
 
-	"github.com/mongodb/grip/recovery"
 	"github.com/pkg/errors"
 )
 
@@ -26,7 +26,12 @@ func (o *oomTrackerImpl) Clear(ctx context.Context) error {
 }
 
 func (o *oomTrackerImpl) Check(ctx context.Context) error {
-	wasOOMKilled, pids, err := analyzeLogs(ctx)
+	analyzer := logAnalyzer{
+		cmdArgs:        []string{"log", "show"},
+		lineHasOOMKill: logContainsOOMKill,
+		extractPid:     getPidFromLog,
+	}
+	wasOOMKilled, pids, err := analyzer.analyzeKernelLog(ctx)
 	if err != nil {
 		return errors.Wrap(err, "error searching log")
 	}
@@ -35,56 +40,19 @@ func (o *oomTrackerImpl) Check(ctx context.Context) error {
 	return nil
 }
 
-func analyzeLogs(ctx context.Context) (bool, []int, error) {
-	var cmd *exec.Cmd
-	wasOOMKilled := false
-	errs := make(chan error)
-	sudo, err := isSudo(ctx)
+func getPidFromLog(line string) (int, bool) {
+	r := regexp.MustCompile(`pid (\d+)`)
+	matches := r.FindStringSubmatch(line)
+	if len(matches) != 2 {
+		return 0, false
+	}
+	pid, err := strconv.Atoi(matches[1])
 	if err != nil {
-		return false, nil, errors.Wrap(err, "error checking sudo")
+		return 0, false
 	}
+	return pid, true
+}
 
-	if sudo {
-		cmd = exec.CommandContext(ctx, "sudo", "log", "show")
-	} else {
-		cmd = exec.CommandContext(ctx, "log", "show")
-	}
-	cmdReader, err := cmd.StdoutPipe()
-	if err != nil {
-		return false, nil, errors.Wrap(err, "error creating StdoutPipe for log command")
-	}
-
-	scanner := bufio.NewScanner(cmdReader)
-	if err = cmd.Start(); err != nil {
-		return false, nil, errors.Wrap(err, "Error starting log command")
-	}
-
-	go func() {
-		defer recovery.LogStackTraceAndContinue("log analysis")
-		select {
-		case <-ctx.Done():
-			return
-		case errs <- cmd.Wait():
-			return
-		}
-	}()
-
-	pids := []int{}
-	for scanner.Scan() {
-		line := scanner.Text()
-		if strings.Contains(line, "low swap") {
-			wasOOMKilled = true
-			if pid, hasPid := getPidFromLog(line); hasPid {
-				pids = append(pids, pid)
-			}
-		}
-	}
-
-	select {
-	case <-ctx.Done():
-		return false, nil, errors.New("request cancelled")
-	case err = <-errs:
-		return wasOOMKilled, pids, errors.Wrap(err, "Error waiting for dmesg command")
-
-	}
+func logContainsOOMKill(line string) bool {
+	return strings.Contains(line, "low swap")
 }
