+++ oom_tracker_linux.go
@@ -3,9 +3,11 @@
 package jasper
 
 import (
-	"bufio"
 	"context"
 	"os/exec"
+	"regexp"
+	"strconv"
+	"strings"
 
 	"github.com/pkg/errors"
 )
@@ -24,7 +26,12 @@ func (o *oomTrackerImpl) Clear(ctx context.Context) error {
 }
 
 func (o *oomTrackerImpl) Check(ctx context.Context) error {
-	wasOOMKilled, pids, err := analyzeDmesg(ctx)
+	analyzer := logAnalyzer{
+		cmdArgs:        []string{"dmesg"},
+		lineHasOOMKill: dmesgContainsOOMKill,
+		extractPid:     getPidFromDmesg,
+	}
+	wasOOMKilled, pids, err := analyzer.analyzeKernelLog(ctx)
 	if err != nil {
 		return errors.Wrap(err, "error searching log")
 	}
@@ -33,46 +40,21 @@ func (o *oomTrackerImpl) Check(ctx context.Context) error {
 	return nil
 }
 
-func analyzeDmesg(ctx context.Context) (bool, []int, error) {
-	var cmd *exec.Cmd
-	wasOOMKilled := false
-	errs := make(chan error)
+func dmesgContainsOOMKill(line string) bool {
+	return strings.Contains(line, "Out of memory") ||
+		strings.Contains(line, "Killed process") || strings.Contains(line, "OOM killer") ||
+		strings.Contains(line, "OOM-killer")
+}
 
-	sudo, err := isSudo(ctx)
+func getPidFromDmesg(line string) (int, bool) {
+	r := regexp.MustCompile(`Killed process (\d+)`)
+	matches := r.FindStringSubmatch(line)
+	if len(matches) != 2 {
+		return 0, false
+	}
+	pid, err := strconv.Atoi(matches[1])
 	if err != nil {
-		return false, nil, errors.Wrap(err, "error checking sudo")
+		return 0, false
 	}
-
-	if sudo {
-		cmd = exec.CommandContext(ctx, "sudo", "dmesg")
-	} else {
-		cmd = exec.CommandContext(ctx, "dmesg")
-	}
-	cmdReader, err := cmd.StdoutPipe()
-	if err != nil {
-		return false, nil, errors.Wrap(err, "error creating StdoutPipe for dmesg command")
-	}
-
-	scanner := bufio.NewScanner(cmdReader)
-	if err = cmd.Start(); err != nil {
-		return false, nil, errors.Wrap(err, "Error starting dmesg command")
-	}
-	pids := []int{}
-	for scanner.Scan() {
-		line := scanner.Text()
-		if dmesgContainsOOMKill(line) {
-			wasOOMKilled = true
-			if pid, hasPid := getPidFromDmesg(line); hasPid {
-				pids = append(pids, pid)
-			}
-		}
-	}
-
-	select {
-	case <-ctx.Done():
-		return false, nil, errors.New("request cancelled")
-	case errs <- cmd.Wait():
-		err = <-errs
-		return wasOOMKilled, pids, errors.Wrap(err, "Error waiting for dmesg command")
-	}
+	return pid, true
 }
